{{cookiecutter.project_name}}/requirements.txt
marshmallow>=3
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
apispec[yaml]
apispec-webframeworks
tox
{%- if cookiecutter.use_celery == "yes" %}
celery[redis]
{%- endif %}
{%- if cookiecutter.wsgi_server == "uwsgi" %}
uwsgi
{%- endif %}
{%- if cookiecutter.wsgi_server == "gunicorn" %}
gunicorn
{%- endif %}
pymysql
cryptography